+++ CHANGELOG.md
@@ -35,7 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added TextColorArgument to minecraft-extras
  - Added LocationArgument to cloud-bukkit
  - Added ServerArgument to cloud-velocity
- - Added LockableCommandManager to cloud-core
  - Added VelocityCommandPreprocessor to cloud-velocity
  - Added PlayerArgument to cloud-bungee
  - Added ServerArgument to cloud-bungee
@@ -44,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added BungeeCommandPreprocessor to cloud-bungee
  - Added named suggestion providers
  - Added a PircBotX implementation
+ - Added registration state to command managers
 
 ### Changed
  - Allow for combined presence flags, such that `-a -b -c` is equivalent to `-abc`